requirements.txt
fake-useragent>=2.2.0
httpx>=0.28.0
hyperliquid-python-sdk>=0.20.0
loguru>=0.7.0
python-dotenv>=1.1.0
web3>=7.13.0